+++ tests/test_festival2_list.py
@@ -167,6 +167,10 @@ class PeriodTestCase(unittest.TestCase):
         self.assertEqual(date(2020, 5, 1), sd2)
         self.assertEqual(date(2020, 5, 31), ed2)
 
+        sd3, ed3 = Period.solar_year(2020, 2021)
+        self.assertEqual(date(2020, 1, 1), sd3)
+        self.assertEqual(date(2021, 12, 31), ed3)
+
     def test_lunar_period(self):
         sd1, ed1 = Period.lunar_year(2020)
         self.assertEqual(LunarDate(2020, 1, 1), sd1)
@@ -187,6 +191,10 @@ class PeriodTestCase(unittest.TestCase):
         sd5, ed5 = Period.lunar_month(2020, 5, leap=0)
         self.assertEqual(LunarDate(2020, 5, 1, 0), sd5)
         self.assertEqual(LunarDate(2020, 5, 30, 0), ed5)
+
+        sd6, ed6 = Period.lunar_year(2020, 2021)
+        self.assertEqual(LunarDate(2020, 1, 1), sd6)
+        self.assertEqual(LunarDate(2021, 12, 29), ed6)
 
 
 class WrappedDateTestCase(unittest.TestCase):